third_party/blink/web_tests/platform/generic/crypto/subtle/pbkdf2/deriveBits-rfc6070-test-vectors-expected.txt

Tests importKey/deriveBits for PBKDF2

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "PBKDF2"
PASS tmpKey.usages.join(',') is "deriveKey,deriveBits"
PASS: deriveBits should be [0c60c80f961f0e71f3a9b524af6012062fe037a6] and was
PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "PBKDF2"
PASS tmpKey.usages.join(',') is "deriveKey,deriveBits"
PASS: deriveBits should be [ea6c014dc72d6f8ccd1ed92ace1d41f0d8de8957] and was
PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "PBKDF2"
PASS tmpKey.usages.join(',') is "deriveKey,deriveBits"
PASS: deriveBits should be [4b007901b765489abead49d926f721d065a429c1] and was
PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "PBKDF2"
PASS tmpKey.usages.join(',') is "deriveKey,deriveBits"
PASS: deriveBits should be [3d2eec4fe41c849b80c8d83662c0e44a8b291a964cf2f07038] and was
PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "PBKDF2"
PASS tmpKey.usages.join(',') is "deriveKey,deriveBits"
PASS: deriveBits should be [56fa6aa75548099dcc37d7f03425e0c3] and was
PASS successfullyParsed is true

TEST COMPLETE
